Create 2021-8-13-QS1.md
---
layout: post
mathjax: true
title: The Quadratic Sieve algorithm for Integer Factorization

---

{% include mathjax.html %}  

The Quadratic Sieve is the second fastest algorithm for factoring large semiprimes. It's the fastest for factoring ones which are lesser than 100 digits long.   

Some Background first. 

Most commonly used factorization methods today rely on a simple mathematical Identity    

$X^2 - Y^2 = (X+Y)(X-Y)$

**Fermat's Factorization**  

To factor $N$ we look for an integer $b$ such that $N + b^2$ is a perfect square.

i.e. $N^2 + b^2 = a^2$  

So $N = a^2 - b^2$

$N = (a + b)(a - b)$  

So we have factored $N$.

For e.g. Let $N = 8051$   

Try $8051 + 1^2, 8051 + 2^2, 8051 + 3^2 .....$, till we hit

$8051 + 7^2 = 90^2$

So $8051 = (90 + 7)(90 - 7)$   

$= 83$ x $97$

$\sqrt(8051) = 89.73$ Both are factors are quite close to the square root.

The further away the 2 factors of $N$ are away from $\sqrt(N)$, the more difficult this method becomes. Only a small fraction of numbers have a factor close to their square root.

Let's take $N = 799$. It will take us trying from 1 to 15, till we find $799 + 15^2 = 32^2$

So $799 = 17 * 47$

**Kraitchik's improvement** 

Kraitchik reasoned that instead of trying to find a & b such that $a^2 - b^2$, it may be sufficient to find ones such that it's equal to a multiple of N.

For e.g. try with $3 * 799$ instead of 799.

$3 * 799 = 2397$  

Try $2397 + 1^2, 2397 + 2^2$ etc.   

We get a square right at the 2nd try

$2397 + 2^2 = 49^2$

A small change in this method is we now need $gcd(N, a + b)$ & $gcd(N, a - b)$ to find the factors

$gcd(799, 49 + 2) = 17$

So we get $799 = 17 * 47$   

What we did here is $kN + b^2 = a^2$  

This can also be represented as a Congruence i.e. 

$a^2 \equiv b^2 \bmod N$  

So the same thing could be done starting with $ceil(\sqrt 799) = 29$ 

$29^2 \bmod 799$ - not a square 

$30^2 \bmod 799$ - not a square 

$31^2 \bmod 799$ - not a square 

$32^2 \bmod 799 \equiv 225 = 15^2$

$799 = (32 + 15) . (32 - 15) = 47 * 17$

**Dixon's Factorization**   

Dixon replaced the condition "is a square" with a much weaker condition "is a product of small prime powers". These are easier to find. Once you find several of them, you then combine them by multiplication to find a combination which is a square. 

$N = 63787$   

$439^2 \bmod 63787 \equiv 1360 = 2^4 * 5 * 17$  

$445^2 \bmod 63787 \equiv 6664 = 2^3 * 7^2 * 17$

$449^2 \bmod 63787 \equiv 10240 = 2^{11} * 5$

Multiply the above 3 congruences  


$(439^2 * 445^2 * 449^2) \equiv (2^{18} * 5^2 * 7^2 * 17^2) \bmod 63787$

$7270^2 = 49492^2 \bmod 64787$

$gcd(63787, 49492 + 7270) = 281$

$63787 = 281 * 227$  

The Factorisation is complete.   

Let's summarize this method to factorize $N$

1. Find many integers such that they are a Quadratic Residue $\bmod N$. Find many $c_i$ such that $c_i \equiv {a_i}^2 \bmod N$.   

2. Take a product of some of the $c_i$'s such that their product is a product of even prime powers. i.e ${c_i}_1 . {c_i}_2 . {c_i}_3 ... \equiv b^2 \bmod N$. 

3. Use $gcd(N, a - b)$ to find the factors of $N$. There is a reasonable chance that the gcd is a factor of $N$   


In Step 1 we look for Quadratic residues $c_i$'s which are B-Smooth. To factor $N$, we can calculate an approximate value of $B$ using the formula  

$L =  e^{\sqrt {\ln(N)ln(ln(N))}}$

$B \approx L^{\frac {1}{\sqrt 2}}$

If we want to factor $N = 87463$, we calculate $B \approx 42$. So we are looking for 43-Smooth Quadratic Residues $\bmod N$. 

The first question is how do we achieve Step 1 - find many 43-smooth integers such that they are a Quadratic Residue $\bmod N$. Enter the Quadratic Sieve Algorithm. If you are unfamiliar with what 'Sieving' in general is, then you can lookup the 'Sieve of Eratosthenes'.

**Quadratic Sieve**   

We use a Quadratic Polynomial to do the sieving - hence the name Quadratic Sieve.  

One such polynomial is $F(T) = T^2 - N$. We start with a $T$ which is slightly larger than $\sqrt N$ 

$\sqrt 87463 = 295.74$. So we start with a = 296.  

We look at the list of numbers $F(a), F(a+1), F(a+2) ...$ to sieve    

$F(a) = 296^2 - N = 153$, $F(a+1) = 297^2 - N = 746$, $F(a+2) = 298^2 - N = 1341$ and so on & so forth. 

Our list $LS$ is now $LS = {153, 746, 1341, 1938, 2537, 3138, 3741, ....}$

We try to find 43-smooth numbers in this list by sieving away by prime numbers smaller than or equal to 43 and checking which numbers in the list get sieved all the way down to 1.

First we build what is called as the Factor Base - this is a list of all primes lesser than or equal to B.

{2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41 ,43}

However, there may be some numbers in our list which are not divisible by any of the primes in the factor base. So we can remove them from the Factor Base. Euler's Criteria for Quadratic Residues can be used to check if N is a Quadratic Residue for any of the odd primes.

If $N^{\frac {p-1}{2}} \bmod p \ne 1$, then $N$ is not a QR for that odd prime, hence that prime is not going to divide any of the numbers in our list. For 2, which is an even prime, we can check manually.

Using the above, we can reduce the factor base to 

FB = {2, 3, 13, 17, 19, 29, 41, 43} - so these are all the primes we are going to sieve our list with.   

Our List which we already generated is 

$LS = {153, 746, 1341, 1938, 2537, 3138, 3741, ....}$  

We start with 2. We take each number in the List LS & divide it each number in the Factor Base FB.

The First number in the list is 153. We cannot sieve it by 2 because it's not divisible by 2. Next we try with the next prime in the FB, i.e. 3 - it's divisible by 3 & gives 51, 51 again is divisible by 3 giving 17 - i.e. the original number got divided by 3 twice, i.e. one of it's factors is $3^2$. Next number in the FB is 13, but 17 is not divisible by 13, so we move on to the next prime in the FB, i.e. 17. 17 divides 17 & give us 1. When we reach one, it means we have found a Smooth number.  

$153 = 3^2 * 17$  

One thing to notice here is that checking whether a number is divisible by the prime by using trial division may be expensive. We can avoid this by using the roots to figure out which numbers in the list are going to be divisible by a particular prime.  

For e.g. for the prime $p_i$, the roots of $N \bmod p_i$ will help figure out. The roots can be found using the Tonelli–Shanks algorithm

Let's try it for $p_i = 13$. I am using sagemath here to find the roots of $N \bmod 13$

sage: Mod(87463, 13).sqrt(all=True)

[5, 8]

This gives 2 roots - 5 & 8. Mod (5 - 296, 13) = 8 & Mod(8-296, 13) = 11. 

In our list LS if we start the indexing with 0, this means that the 8th & 11th number will be divisible by 13.  Also, 8+13th, 8+26th, 8+39th etc & 11+13th, 11+26th, 11+39th etc number will be divisible by 13. This can be used to avoid trial division & only sieve numbers at that position.   



[Excel Sheet](https://1drv.ms/x/s!ApBR9uU7g-hGakHn6-v7WVllgbo?e=mTWTNg)